Backport of unified retry / auto-retry filter support (#2020)
Motivation:

Use of `AutoRetry` alongside a `RetryingHttpRequesterFilter` creates issues with the retrying logic exponentially increasing to the product of the two max-retry settings. The conflict arises when both solutions are configured to retry `RetryableException`s which is the case by default.

Modifications:

- `AutoRetry` logic is deprecated, to be removed in the future.
- The logic of the `AutoRetry` is merged with the `RetryingHttpRequesterFilter` which now allows retryable flows. Different exceptions can have different retry, as well as support for retries of responses.

Results:

Unified API for retries.

Note:

This is a cherry-picked combination of 3 PRs to the `main` branch, see  #1905, #2004, #2018.