Fix filters for MultiAddressUrlHttpClientBuilder (#2071)
Motivation:

Recently the API of filters changed to align better with the way
`ExecutionStrategy` should be set. A temporary additional filter
bridging calls via new API to old API was created to assure compatibility.
Unfortunately, the `MultiAddressUrlHttpClientBuilder` was not updated
and legacy filters may not be exercised.

Modifications:

- Properly appending the `NewToDeprecatedFilter` to filters added via
  `MultiAddressUrlHttpClientBuilder`,
- Legacy `StrategyInfluencerAwareConversions#toMultiAddressClientFactory`
  was also updated to include the bridge,
- `MixedFiltersTest` include a case validating the
  `MultiAddressUrlHttpClientBuilder` behavior.

Result:

Deprecated methods on filters are properly executed when using
`MultiAddressUrlHttpClientBuilder`.