Preparing for 0.42.3-SNAPSHOT development
Motivation:

Until we have any breaking changes in the `main` branch, it will be
easier to continue 0.42.x development from the `main` to avoid a need
for cherry-picking all commits into `0.42`.
At this point of git history, the `main` and `0.42` branches are
identical except antora-related files which have to be like in the
`main` branch.