Revert "Removed GrpcClientBuilder#MultiClientBuilder (#1809)" (#2127)
Revert "Removed GrpcClientBuilder#MultiClientBuilder (#1809)

This reverts commit 74ac3902a0ba09c72fde79923906ad1c3ac394ee.

Motivation:

`GrpcClientBuilder#MultiClientBuilder` was previously removed, but it
was the only practical way to share the underlying transport between
different service API clients exposed by the same backend. At the same
time, we don't want to expose ways to interact with the internal assembly
of the transport layer, as we attach default filters and don't want to
expose that internal implementation to users. Therefore, we're bringing
the `MultiClientBuilder` back.

Modifications:

- Reverted `GrpcClientBuilder#MultiClientBuilder`,
- Improved Javadoc,
- Added `MultiClientTest` validating the behaviour.

Result:

`GrpcClientBuilder#MultiClientBuilder` can continue to be used by users
migrating from 0.41 release and its' intention is clarified.