Deprecated `PartitionedHttpClientBuilder` and partition client APIs
Motivation:

`PartitionedHttpClientBuilder` is hard to be used because it requires
implementation of a special `ServiceDiscoverer` that we do not provide
OOTB. We are unaware of anyone using this feature, but if there are any
users should consider `ClientGroup` as an alternative or reach out to
us to provide more context why the `PartitionedHttpClientBuilder` is
required.

Modifications:

- Deprecate `PartitionedHttpClientBuilder` and
`HttpClients.forPartitionedAddress`;
- Deprecate everything in `io.servicetalk.client.api.partition` package;

Result:

`PartitionedHttpClientBuilder` is deprecated and can be removed in
future releases.