Add `publishOn`/`subscribeOn` `Executor` overload (#2155)
* Add `publishOn`/`subscribeOn` `Executor` overload
Motivation:
The signature of `publishOn(Executor)` and `subscribeOn(Executor)`
methods changed between 0.41 and 0.42 from
`io.servicetalk.concurrent.api.Executor` to the more general parent
type `io.servicetalk.concurrent.api.Executor`. This is a source
compatible change since the new type is a parent of the old type, but
is not a binary compatible change; code compiled against 0.41 will not
run with 0.42 as it will not find the method it expects with the
`io.servicetalk.concurrent.api.Executor` signature.
Modifications:
Add overloads `publishOn(Executor)` and `subscribeOn(Executor)` for
`io.servicetalk.concurrent.Executor`. For binary compatibility with
0.42 it will also be necessary to cast executors to
`io.servicetalk.concurrent.api.Executor` to ensure that the correct
overload is used.
Result:
Improved compatibility with ServiceTalk 0.42