Use filter strategy for streaming async server (#2156)
Motivation:
Currently the filter strategy is not considered in the computation of
the effective strategy for async streaming servers. This was done
because with a default builder strategy the result could be less
offloading which would be unexpected for that builder strategy.
Modifications:
Consider filter strategy in computing effective strategy for
async streaming servers.
Result:
More predictable offloading behavior.