`RoundRobinLoadBalancer`: increase visibility into health-checking (#2182)
Motivation:

Currently, health-checking events are logged at `DEBUG` level and the
failure cause is not propagated. Users have to change the logging level
and reproduce in order to understand the root cause.

Modifications:

- Log each failed attempt to create a connection at `INFO` level for
visibility;
- Log at `WARN` level when the health-checking is triggered;
- Retain the last cause that triggered health-checking inside
`HealthCheck` state to use it for `toString()`;

Result:

Users see when a connectivity issue occurs, cause is logged and
propagated for subsequent selection attempts if there are no other
`ACTIVE` hosts.