Let `ConnectionFactory` see difference between proxied and direct connections (#2169)
Motivation:

Connection initialization sequence is different for a secure HTTP proxy
tunnels.  It's necessary for the `ConnectionFactory` and
`TransportObserver` to know if the `newConnection` is invoked for a
regular connection or a proxy tunnel in order to correctly count timing
of events.

Modifications:

- Introduce new `HttpContextKeys#HTTP_TARGET_ADDRESS_BEHIND_PROXY` key;
- Enhance `ProxyConnectConnectionFactoryFilter` and
`AbsoluteAddressHttpRequesterFilter` to fill the value for
`HTTP_TARGET_ADDRESS_BEHIND_PROXY`;
- Enhance proxy tests to verify that the key value always present;

Result:

Users can distinguish attempts to open direct connections and
connections through a proxy.