fix: servicetalk-data-jackson dependency scopes (#2198)
Motivation:
The `servicetalk-data-jackson-jersey` module exposes types that are
located in the `servicetalk-data-jackson` module. Similarly, the
`servicetalk-data-jackson` module exposes types that are in the
`servicetalk-serializer-api` and `servicetalk-serialization-api` modules.
Dependencies that contain types exposed in a module's API should be imported using `api` scope but currently these dependencies are imported using only `implementation` scope. Attempting to use the types results in a `ClassNotFoundException`.

Modifications:
- Change `servicetalk-data-jackson` to an `api` dependency of `servicetalk-data-jackson-jersey`
- Change `servicetalk-serializer-api` and `servicetalk-serialization-api` to `api` dependencies of `servicetalk-data-jackson`

Result:
Consumers of the `servicetalk-data-jackson-jersey` and `servicetalk-data-jackson` modules will be able to reference all types exposed in these modules' APIs without class not found errors or the need to explicitly import `servicetalk-data-jackson`, `servicetalk-serializer-api` or `servicetalk-serialization-api` modules.