Tighter dependency for protoc plugin (#2209)
Motivation:
Currently the `servicetalk-grpc-protoc` module exposes javapoet and
protobuf-java as dependencies though they are used only in the protoc
plugin and are not accurately dependencies of the module.
Modifications:
Use `compileOnly` for protoc plugin dependencies and change shadowJar
to build using `compileClasspath` configuration. The dependencies are
also needed for `testImplementation`.
Result:
Fewer exposed dependencies, especially dependency which is not
actually used by module outside of protoc plugin.