Add dependencies BOM (#2210)
* Add dependencies BOM
Motivation:
The current ServiceTalk BOM file includes only ServiceTalk modules and
does not include version information for the external dependencies
referenced by ServiceTalk. This information can be found transitively
but it is beneficial to be explicitly include it all in a BOM file.
Modifications:
Add an additional BOM `servicetalk-dependencies` which includes both
ServiceTalk modules and BOMs and versions for external dependencies.
Result:
An additional ServiceTalk BOM which can improve dependency version
resolution by explicitly identifying the preferred and tested
versions of external components which ServiceTalk utilizes.