Use `servicetalk-dependencies` BOM internally (#2217)
Motivation:
The new `dependencies` BOM introduced in #2210 can be used to improve
the management of external dependencies by reducing use of explicit
versions within the module `build.gradle` files.
Modification:
Use `dependencies` BOM for every module which includes `api` or
`implementation` external dependencies. The BOM is imported as `api`
platform only in cases where the external dependency is also imported
`api`. `implementation` platform is used in cases where the external
dependencies are only `implementation` dependencies themselves.
The Jackson-BOM is no longer used because of issues with how it is
versioned and our desire for a narrower dependency on
`jackson-databind` only.
Result:
Fewer explicit version specifications and better managed dependency
versioning.