Migrate `ON_CONNECTION_SELECTED_CONSUMER` to request context
Motivation:

`AbstractLifecycleObserverHttpFilter` uses a secret
`ON_CONNECTION_SELECTED_CONSUMER` context key to capture the selected
connection down the processing chain. Instead of relying on
`AsyncContext` that can be disabled, we can use a context of the request.

Modifications:

- Store `ON_CONNECTION_SELECTED_CONSUMER` in `request.context()` instead
of `AsyncContext`;
- Use `Runnable` to clear the key on exchange finally;

Result:

`AbstractLifecycleObserverHttpFilter` will report `onConnectionSelected`
even if `AsyncContext` is disabled.