Fix regression of client `offloadNone()` strategy (#2192)
Motivation:
The changes in #2108 introduced a regression in the handling of
computation of the effective strategy of a client when a
strategy `offloadNone()` (or the deprecated `offloadNever()`) was used
with the  `HttpClientBuilder`. Rather than preventing all offloading,
the computed offloading was used.

Modifications:
If the strategy specified on the `HttpClientBuilder` does not specify
any offloads then no offloading will be used.

Result:
The offloading behavior matches the requested behavior specified on
the  `HttpClientBuilder`.