Cancellation can leave an HTTP/1.x connection in half-closed state
Motivation:

If a response is cancelled at the connection level (skipping explicit
closure of the connection in `LoadBalancedStreamingHttpClient`), it will
hang in half-closed state (inbound closed) if the server doesn't proceed
with sending more bytes on that connection (outbound will never be
closed). Also, cancellation marks the connection as
`CHANNEL_CLOSED_INBOUND`, which is misleading because the closure is
initiated by the client.

Modifications:

- Enhance `ResponseCancelTest` to verify the described scenario;
- Adjust `NettyChannelPublisher` to close both outbound (first, to
correctly mark the connection close reason) and inbound;

Result:

Cancelled HTTP/1.x connection is always closed correctly.