Handle HTTP/1.1 response cancelation same way on all levels
Motivation:

When response cancel is generated at client level,
`LoadBalancedStreamingHttpClient` has a logic in place to close the
connection to mitigate a race between connection availability and
closure. However, this logic is skipped if cancel is generated from
connection-level filter.

Modifications:

- Move connection closure from `LoadBalancedStreamingHttpClient` to
`PipelinedStreamingHttpConnection`;
- Adjust tests and comments;

Result:

HTTP/1.1 connection is always closed in case of cancellation. We do that
from the caller thread to proactively set `onClosing` state.