`TimeoutHttpRequesterFilter` as a connection filter causes `ClosedChannelException` (#2263)
Motivation:

If applied as a connection-level filter, it skips the logic defined in
`LoadBalancedStreamingHttpClient` for the cancellation. Cancel will be
scheduled on the event-loop, then `onError` is propagated.
`LoadBalancedStreamingHttpClient` will mark the request as finished, it
won’t see cancel. As the result, the same connection can be selected for
another request, and users see `ClosedChannelException`.

Modifications:

- For `StreamingHttpConnectionFilterFactory` variant used for HTTP/1.X
connections, apply `BeforeFinallyHttpOperator` with `cancel` callback
similar to `LoadBalancedStreamingHttpClient`;
- Enhance mocks to account for new behavior;
- Move the common logic of extracting a fallback executor as a static
method in `AbstractTimeoutHttpFilter`;

Result:

HTTP/1.X requests don't fail with `ClosedChannelException` if a previous
request on the same connection times out.