Add `.p12` stores for tests which require format (#2286)
Motivation:
Some new tests require certificates and keys to be presented in PKCS#12
keystore format.
Modifications:
Add PKCS#12 stores for client, server and truststore. Certs are
regenerated as server CA cert was expired and could not be used for
signing of localhost_server server cert. Key size is updated to 3072
bits, the current FIPS 140-2 SP#800 requirement for RSA keys, as this
size is what will typically be required for commercial use.
Result:
More test resources