Handle HTTP/1.1 response cancelation same way at all levels (#2266)
Motivation:

When response cancel is generated at the client level,
`LoadBalancedStreamingHttpClient` has a logic in place to close the
connection to mitigate a race between connection availability and
closure. However, this logic is skipped if cancellation is originated from
a connection-level filter.

Modifications:

- Move cancellation handling from `LoadBalancedStreamingHttpClient` to
`AbstractStreamingHttpConnection`;
- Propagate `OnStreamClosedRunnable` using `request.context()`;
- Remove `StreamingHttpRequestWithContext`;
- Remove "cancel handling" logic from `TimeoutHttpRequesterFilter`
introduced in #2263;
- Add more comments and javadoc;
- Adjust tests and comments;

Result:

Request to cancel is intercepted after all user-defined filters to make sure
we catch all sources of cancellation.