Hide use of `sun.misc.Unsafe` from `javac` (#2303)
Motivation:
Attempting to build using a Java 9+ `javac` with the `--release`
option currently fails because the ServiceTalk source contains references to
the `sun.misc.Unsafe` class. This class is optionally used by ServiceTalk
but the appearance of the class in the source currently requires that 
source be compiled with Java 8 or without the `--release` option to avoid
compilation failure due to 
[JDK-8214165](https://bugs.openjdk.org/browse/JDK-8214165).
Modifications:
Use reflection and method handles rather than direct class use for
`sun.misc.Unsafe`.
Result:
ServiceTalk may be compiled using the `javac --release` option.