`AsciiBuffer.indexOf` returns incorrect result for 2-bytes characters (#2352)
* `AsciiBuffer.indexOf` returns incorrect result for 2-bytes characters

Motivation:

`AsciiBuffer.indexOf` casts `char` to `byte` by dropping the higher 8
bits. As the result, it may return an index of an ascii character that
has the same lower 8 bits.

Modifications:

- Check character length and return `-1` if it's 2-bytes character;

Result:

Correct result when users try to find an index of a 2-bytes character in
`AsciiBuffer`. Consistency with other `CharSequence` implementations,
like `String` and `StringBuilder`.