Racy `cancel` from transport does not always cancel the message body
Motivation:

When `cancel` from the transport races with HTTP message completion, we
can receive `cancel` after `Single<HttpMetaData>` completes but before
we subscribed to the `messageBody` publisher. In this case, we won't
subscribe to the `messageBody` because we use `concat` operator. This
does not allow `BeforeFinallyHttpOperator` to observe a terminal event.
Race is possible when offloading is enabled. `cancel` comes from an IO
thread, HTTP meta-data completes on offloaded thread.

Modifications:

- For every case when we concat a completed source with the message body
add `afterFinally` to guarantee propagation of the cancel for the entire
flat HTTP stream;

Result:

`HttpLifecycleObserverTest.testClientCancelsRequestBeforeResponse()`
does not race anymore.