Server control flow should close connection on `DecoderException`
Motivation:

#2367 changed server-side control flow. After that, `connection.write`
doesn't start before we received a request meta-data. As the result, all
errors that happen during parsing of the meta-data do not reach
`WriteStreamSubscriber` and can leave a connection in undefined state,
when we logged an error but didn't clean up the channel state.

Modifications:

- Close the channel inside `ErrorLoggingHttpSubscriber` if it was not
closed already. It covers a case when an error occurs on the `Single` of
request meta-data;
- Add tests to verify the expected behavior;

Result:

Server channel is closed when an early request error happens, client
doesn't hang forever.