RoundRobinLoadBalancer fix connection leak during graceful closure
Motivation:
RoundRobinLoadBalancer clears state of all hosts and each host clears
all connections when closing. However a common pattern is to initiate
gracefulClosure with a timeout, and then force a full close after a
timeout interval. In this scenario RRLB has discarded the reference to
connections and will not full close the connections.

Modifications:
- asyncCloseable needs to keep state about what hosts are being closed
  and each host can't dereference connections until they are fully closed.
- switch back to using onClose to cleanup the connection list as otherwise
  we may leak connections that have started graceful closure and never finished.