Provide better exception when connection limit reached in LimitingCon… (#2454)
Motivation:

Before this change, when the LimitingConnectionFactoryFilter does not allow more connections
for a given host, a generic ConnectException is returned. Developer productivity can be
improved since at the moment the user needs to inspect the text message for further
information and to distinguish different failure cases.

Modifications:

This changeset subclasses the ConnectException for backwards compatibility and returns a
ConnectionLimitReachedException instead. The actual message is also preserved for
compatibility reasons.

Result:

Better user experience when matching on the exception instance while still being
backwards compatible.